requirements.txt
requests
pillow
colorthief
python-dotenv